+++ packages/core/src/components/EmptyState/EmptyState.types.ts
@@ -1,7 +1,7 @@
 import type React from "react";
 import { type LinkProps } from "../Link/Link";
 import { type VibeComponentProps } from "src/types";
-import { type ButtonProps, type ButtonType } from "../Button";
+import { type ButtonProps, type ButtonType } from "@vibe/button";
 
 export type EmptyStateLayout = "default" | "compact";