+++ packages/core/src/components/ListItem/ListItem.tsx
@@ -19,6 +19,7 @@ import useMergeRef from "../../hooks/useMergeRef";
 import { ListContext } from "../List/utils/ListContext";
 import { ListItemComponentType as ListItemComponentTypeEnum } from "./ListItemConstants";
 import { type ListItemElement, type ListItemSize } from "./ListItem.types";
+import { type TooltipProps } from "../Tooltip/Tooltip";
 import styles from "./ListItem.module.scss";
 
 export interface ListItemProps extends VibeComponentProps {
@@ -62,6 +63,10 @@ export interface ListItemProps extends VibeComponentProps {
    * The ARIA role of the list item.
    */
   role?: AriaRole;
+  /**
+   * Props passed to the tooltip displayed when hovering over the text.
+   */
+  tooltipProps?: Partial<TooltipProps>;
 }
 
 const ListItem = forwardRef(
@@ -79,7 +84,8 @@ const ListItem = forwardRef(
       children,
       "aria-current": ariaCurrent,
       "data-testid": dataTestId,
-      role = "option"
+      role = "option",
+      tooltipProps
     }: ListItemProps,
     ref: React.ForwardedRef<HTMLElement>
   ) => {
@@ -134,6 +140,7 @@ const ListItem = forwardRef(
         role={role}
         tabIndex={tabIndex}
         aria-current={ariaCurrent}
+        tooltipProps={tooltipProps}
       >
         {children}
       </Text>