+++ packages/core/src/components/Toast/ToastButton/ToastButton.tsx
@@ -1,4 +1,4 @@
-import Button, { type ButtonProps } from "../../Button/Button";
+import { Button, type ButtonProps } from "@vibe/button";
 import React, { type FC } from "react";
 import { ComponentDefaultTestId } from "../../../tests/constants";
 import { getTestId } from "../../../tests/test-ids-utils";