+++ packages/docs/src/layout/related-components/descriptions/slider-description.tsx
@@ -1,0 +1,16 @@
+import React, { useMemo } from "react";
+import { RelatedComponent } from "vibe-storybook-components";
+import { Slider } from "@vibe/core";
+
+export const SliderDescription = () => {
+  const component = useMemo(() => <Slider />, []);
+
+  return (
+    <RelatedComponent
+      component={component}
+      title="Slider"
+      href="/?path=/docs/components-slider--docs"
+      description="Displays a slider bar with range selection"
+    />
+  );
+};