+++ packages/docs/src/pages/hooks/useClickableProps/useClickableProps.stories.tsx
@@ -1,0 +1,33 @@
+import React, { useCallback, useRef } from "react";
+import { useClickableProps } from "@vibe/core";
+import "./useClickableProps.stories.scss";
+
+export default {
+  title: "Hooks/useClickableProps"
+};
+
+export const Overview = {
+  render: () => {
+    const ref = useRef<HTMLDivElement>(null);
+    const onClick = useCallback(() => alert("click!"), []);
+
+    const clickableProps = useClickableProps(
+      {
+        onClick: onClick,
+        id: "clickable-id",
+        ariaHidden: false,
+        ariaHasPopup: false,
+        ariaExpanded: false
+      },
+      ref
+    );
+
+    return (
+      <div {...clickableProps} className="monday-storybook-use-clickable-props" ref={ref}>
+        I act like a button
+      </div>
+    );
+  },
+
+  name: "Overview"
+};