+++ packages/storybook-config/src/index.ts
@@ -1,2 +1,0 @@
-export { default as createStoryMetaSettings } from "./createStoryMetaSettings";
-export { default as createStoryMetaSettingsDecorator } from "./createStoryMetaSettingsDecorator";