+++ packages/config/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.3](https://github.com/mondaycom/vibe/compare/@vibe/config@3.0.2...@vibe/config@3.0.3) (2025-10-30)
+
+**Note:** Version bump only for package @vibe/config
+
+
+
+
+
+## [3.0.2](https://github.com/mondaycom/vibe/compare/@vibe/config@3.0.1...@vibe/config@3.0.2) (2025-10-25)
+
+**Note:** Version bump only for package @vibe/config
+
+
+
+
+
 ## 3.0.1 (2025-09-14)
 
 **Note:** Version bump only for package @vibe/config